Fix ansible 2.x compat. and modify the default functionality
This update will remove kickstart settings for disabling ipv6, update tasks
to ansible 2.x and add localization variables (keyboard, lang, timezone).
It also adds handlers for httpd and xinetd restart, selinux perms
for /var/www/provision and change the default kickstart packages to only
include @core (minimal install) instead of @base.

The sysctl.conf disable IPv6 on the kernel level -settings are removed
from the kickstart file + the network --noipv6  that is now obsolete.
The kickstart file is now splitted into separate files for packages and
a list of post scripts to include. These files can be set via new
variables.

Add support for hosts file generation from inventory. Off by default

Split tasks into install and configure and fix missing configs

The install and configure tasks are split into to files to be able to
run only the config update tasks by setting pxe_setup: false when
including the role. Also the two configs used by the role
ansible-role-dhcp-kickstart are created now by default to fix the
problem of having include statements in dhcp_common_parameters when
building the dhcp server the first time. These include lines will fail
when restarting the dhcpd before the kickstart role is run.

Role refactored completely

Main changes: groups in the role are gone, kickstart generation has more
parameters and dhcp/pxe is disabled by default as the role is intended to
be used now in the host loop along with delegate_to. The templates don't
use hostvars[] directly anymore and this fixes a lot of quirks with variable
precedence in Ansible.

Fix problem with handlers not running as root

Force become: yes to the handlers as ansible's new include_role is a bit
buggy and can cause the handlers to not be run as root even if the
calling task has become: yes.

Add SUSE autoyast support

This changeset adds support for SLES autoyast bootstrapping. A autoyast file
provided by SUSE (to be as minimal as possible but working in all
situations) is used as a template. The kickstart/autoyast script directory is
renamed to bootstrap (located under /var/www/html).
The variable use_kickstart is true by default so that this role will produce
kickstart files by default if you enable use_pxe for a host.

Add missing enable and start httpd on boot

New task to flag a host for reinstall

Other cleanup so that delgation is hidden from the end user of the role.

Comment to clarify inconsistent handler delegation