Split tasks into install and configure and fix missing configs
The install and configure tasks are split into to files to be able to
run only the config update tasks by setting pxe_setup: false when
including the role. Also the two configs used by the role
ansible-role-dhcp-kickstart are created now by default to fix the
problem of having include statements in dhcp_common_parameters when
building the dhcp server the first time. These include lines will fail
when restarting the dhcpd before the kickstart role is run.