Role refactored completely
Main changes: groups in the role are gone, kickstart generation has more
parameters and dhcp/pxe is disabled by default as the role is intended to
be used now in the host loop along with delegate_to. The templates don't
use hostvars[] directly anymore and this fixes a lot of quirks with variable
precedence in Ansible.