Add SUSE autoyast support
This changeset adds support for SLES autoyast bootstrapping. A autoyast file
provided by SUSE (to be as minimal as possible but working in all
situations) is used as a template. The kickstart/autoyast script directory is
renamed to bootstrap (located under /var/www/html).
The variable use_kickstart is true by default so that this role will produce
kickstart files by default if you enable use_pxe for a host.