Use a memtest.0 found on the Internet
This has been tested and it boots an memtest86+ v5.01 from 2015.

According to a forum post by an mcb30 Lead Developer of iPXE memtest.0

"This is memtest in the form of a PXE NBP:"

I checked in version 7.5 of memtest86+ src.tgz from the ISO and it does
not contain "pxe" anywhere in the code. Presumably one could make a
similar patch to 7.5 to link memtest with pxe.

The patch involves assembly = I am unable to judge if this is safe or
not.