Use a memtest.0 to optionally boot into a memtest instead of kickstart (#10)
This has been tested and it boots a Dell R640 and R740 BIOS (not UEFI) with an memtest86+ v5.01 from 2015.

Controlled by touching a file in memtest86/(mynodes short hostname). The file is removed before booting into memtest which autostarts. After memtest is done the server is waiting for a keystroke (escape) to quit and resume booting.

memtest86 has preference over reinstall (if both files exist the server runs a memtest86). On next boot it will reinstall.

https://www.memtest86.com/features.htm says one has to get the "site edition" of memtest86 for the network (PXE) boot feature.

Some background details about what we are booting (and what needs to be copied onto a web server which the nodes can reach via http within the iPXE script):

According to a forum post by an mcb30 Lead Developer of iPXE memtest.0 is:

" This is memtest in the form of a PXE NBP: it has the advantage that you can return to iPXE afterwards (and report the test result, and potentially capture a screenshot of the memtest output and send it to a reporting server via HTTP POST)."

I checked in version 7.5 of memtest86+ src.tgz from the ISO and it does
not contain "pxe" anywhere in the code. It is possible to make a
similar patch to 7.5's Makefile and add the pxe.S file to build a memtest.0 but with that build the server did not manage to boot into memtest. Maybe it would work if one took in more of the patches from https://git.ipxe.org/people/mcb30/memtest.git/shortlog/refs/heads/pxe501