More kickstart settings and more testing (#17)
Testing

now we actually http get boot.py after the reinstall and the memtest86 files have been touched. This in travis/testing uses the hostname of the container where the test is running.
also, grep the output from those curls to include the things we expect them to include
we also run ksvalidator from the pykickstart RPM on the kickstart files, not only in testing but also when the kickstart gets templated in. Meaning if one were to add something bad via variables that would result in an invalid kickstart config ansible gives an error.
the changes to kickstart.j2 and boot.py are by defining variables in some places in group_vars.
to test that the above works like it should, we now fail the travis build if grep for "sdc" for compute3 doesn't find anything
Lots of more variables one can use to define what will end up in the kickstart. None of them should be required

To pxe_json or the boot.py:

serialport #you can now define serialport configuration during anaconda/kickstarting
backup pxe_nodes.json file if it changes
To kickstart.j2 configs:

The below would need to be set for the group, so for example group_vars/groupname/vars.yml - it's not enough to specify them in group_vars/all.yml because of how we use hostvars[] to write the kickstart templates. More details about this bit added to README.MD

kickstart_grubby_args #use this to define arguments to the default kernel (the one the OS will boot on after the kickstart). This is so one can use IPv6 if so desired or modify the consoles
kickstart_grubby_remove_args # use this to remove arguments from the grub of the default kernel.
yum_proxy # if defined is used also in the url and repo bits
kickstart_lang #if undefined: en_US.UTF-8
kickstart_keyboard # if undefined: fi-latin1
kickstart_timezone # if undefined: Europe/Helsinki
kickstart_log_host # can be used to ship logs to a remote host during kickstart
dhcp_kickstart_manage_packages #if this is defined then kickstart_packages is expected to contain %packages. If it's not defined we keep the backwards compatible list of packages installed and uninstalled as before this PR
kickstart_packages #a string which is a list of packages to install or uninstall
kickstart_extra_pre_option # one can choose interpreter of the extra_pre_commands
kickstart_extra_pre_commands
kickstart_extra_post_commands # only bash for post_commands
As part of kickstart templates we now backup=yes the files, meaning if every time ansible template would change a kickstart file it would store a backup of the old file in the same directory